Module pallet_trust::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

DefaultCurrent

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageCurrentIssued

_GeneratedPrefixForStorageCurrentRequests

_GeneratedPrefixForStorageCurrentRevoked

_GeneratedPrefixForStorageTrustIssuance

_GeneratedPrefixForStorageTrustRequestList

_GeneratedPrefixForStorageTrustRevocation

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Functions

__type_value_for_default_current

Type Definitions

CurrentIssued

The total number of trust actions currently active

CurrentRequests

CurrentRevoked

The current number of _non_trust actions currently active

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

TrustIssuance

A Map of lists of all addresses that each address has issued trust for

TrustRequestList

A map listing all requests for trust from one account to another.

TrustRevocation

A Map of lists of all addresses that each address has issued trust for